requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
yarl==1.9.4
